requirements.txt
streamlit
langchain>=0.2.0
langgraph>=0.1.0
langchain-groq
langchain-community
langchain-core
sentence-transformers
faiss-cpu
easyocr
openai-whisper
torch
sympy
python-dotenv
pydantic
